requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-login==0.6.0
flask-wtf==1.0.0
email-validator==1.1.3
flask-bcrypt==1.0.1
pillow==9.0.0
flask-mail==0.9.1
pyjwt==2.3.0
flask-cors==3.0.10
python-dotenv==0.19.2
gunicorn==20.1.0
itsdangerous==2.0.1
